Fixed offset issue when using LedBarGraph.Percentage with a large number of LEDs
When lighting a percentage of non-pwm lights, 0.5 is added to the percentage.
This is to stop floating point equality errors and works fine for 10 LEDs.
However, if you map 20 or more LEDs into a LedBarGraph, this results in 1 or more LEDs being erroneously turned on.

The fix is to instead add half of the increment, or 0.5 * (1 / LedCount)